+++ pages/calendar/activaciones/adm_activaciones.js
@@ -1,3 +1,5 @@
 import * as registroHoras from './registrohoras.js';
+import * as licencias from '../licencias/licencias.js';
 
-registroHoras.init(new Date(), new Date());
+const user_calendar = await registroHoras.init(new Date(), new Date());
+licencias.init(new Date(), new Date(), user_calendar);